refactor: kotlin -> gradle
