--- build.gradle.kts
@@ -1,17 +1,0 @@
-/*
- * This file was generated by the Gradle 'init' task.
- *
- * This is a general purpose Gradle build.
- * To learn more about Gradle by exploring our Samples at https://docs.gradle.org/8.8/samples
- */
-plugins { java }
-
-repositories { mavenCentral() }
-
-dependencies {
-    testImplementation(platform("org.junit:junit-bom:5.10.2"))
-    testImplementation("org.junit.jupiter:junit-jupiter")
-}
-tasks.test {
-    useJUnitPlatform()
-}
