+++ src/main/java/com/javacards/cards/model/DealerHand.java
@@ -1,4 +1,4 @@
-package main.model;
+package com.javacards.cards.model;
 
 import java.util.*;
 /*
