+++ src/main/java/com/javacards/cards/model/Value.java
@@ -1,4 +1,4 @@
-package main.model;
+package com.javacards.cards.model;
 
 /*
  * Value.java
