--- src/main/java/main/CardApplication.java
@@ -1,7 +1,0 @@
-package main;
-
-public class CardApplication {
-    public static void main(String[] args) {
-        System.out.println("Main Ran");
-    }
-}
